% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bismer.r
\name{insult}
\alias{insult}
\title{Generate a random Shakespearean insult}
\usage{
insult(thou = TRUE)
}
\arguments{
\item{thou}{if \code{TRUE} then the insults are prefixed with "thou"}
}
\description{
Generate a random Shakespearean insult
}
